separation between app and server
